# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: API Seed Desafio CDC
  description: Documentação da API para o projeto Seed Desafio CDC.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de desenvolvimento local
paths:
  /autores:
    post:
      summary: Cadastrar um novo autor
      description: Endpoint para cadastrar um novo autor no sistema.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  example: João Silva
                email:
                  type: string
                  example: joao.silva@email.com
                descricao:
                  type: string
                  example: Autor de livros de ficção científica.
      responses:
        '201':
          description: Autor cadastrado com sucesso
          headers:
            Location:
              description: URL do recurso recém-criado
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Autor'
  /autores/{id}:
    get:
      summary: Consultar os dados de um autor
      description: Endpoint para consultar os dados de um autor específico pelo ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID do autor a ser consultado.
          schema:
            type: integer
      responses:
        '200':
          description: Dados do autor retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Autor'
        '404':
          description: Autor não encontrado
...
components:
  schemas:
    Autor:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: João Silva
        email:
          type: string
          example: joao.silva@email.com
        descricao:
          type: string
          example: Autor de livros de ficção científica.
...